refactor: instution simple dto 에 요양기관 id 추가
