+++ src/main/java/com/becareful/becarefulserver/domain/matching/dto/RecruitmentDto.java
@@ -5,6 +5,8 @@ import static java.util.stream.Collectors.groupingBy;
 import com.becareful.becarefulserver.domain.caregiver.domain.WorkSalaryUnitType;
 import com.becareful.becarefulserver.domain.common.domain.DetailCareType;
 import com.becareful.becarefulserver.domain.matching.domain.Recruitment;
+import com.becareful.becarefulserver.domain.nursing_institution.dto.InstitutionSimpleDto;
+
 import java.time.DayOfWeek;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
