+++ src/main/java/com/becareful/becarefulserver/domain/matching/dto/RecruitmentSimpleDto.java
@@ -3,6 +3,8 @@ package com.becareful.becarefulserver.domain.matching.dto;
 import com.becareful.becarefulserver.domain.caregiver.domain.WorkSalaryUnitType;
 import com.becareful.becarefulserver.domain.common.domain.CareType;
 import com.becareful.becarefulserver.domain.matching.domain.Recruitment;
+import com.becareful.becarefulserver.domain.nursing_institution.dto.InstitutionSimpleDto;
+
 import java.time.DayOfWeek;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
